pyproject.toml
[tool.poetry]
name = "glints-restful"
version = "0.1.0"
description = ""
authors = ["Lance Chang <virtuouslycan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.1"
SQLAlchemy-Utils = "^0.37.8"
Flask-Script = "^2.0.6"
python-dateutil = "^2.8.2"
Python-Deprecated = "^1.1.0"
flask-restx = "^0.5.0"
uWSGI = "^2.0.19"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^5.5"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
